Při schválení žádosti o změnu názvu třídy se aktualizuje i URL
Dříve se URL třídy neaktualizovalo.
Abstraktní funkce NameChangeRequest::approve() je odebrána a je implementována v ClassNameChangeRequest i UserNameChangeRequest
